Fix: use reliable tab-count method
