+++ app/layout.tsx
@@ -32,7 +32,9 @@ const pretendard = localFont({
   variable: '--font-pretendard',
 });
 
-const SessionIdProvider = dynamic(() => import('@/context/SessionIdContext'), { ssr: false, loading: () => (<></>)});
+const SessionIdProvider = dynamic(() => import('@/context/SessionIdContext'), { ssr: false, loading: () => (<></>) });
+
+import ClearStorageDependOnTabs from '@/context/ClearStorageDependOnTabs';
 
 export default function RootLayout({
   children,
@@ -45,11 +47,13 @@ export default function RootLayout({
         <meta name="naver-site-verification" content="dee9637347b73df323379604875290ad69eb2ee8" />
       </head>
       <body className={`${inter.className} ${pretendard.className}`}>
-        <SessionIdProvider>
-          <TopLevelLayout>
-            {children}
-          </TopLevelLayout>
-        </SessionIdProvider>
+        <ClearStorageDependOnTabs>
+          <SessionIdProvider>
+            <TopLevelLayout>
+              {children}
+            </TopLevelLayout>
+            </SessionIdProvider>
+        </ClearStorageDependOnTabs>
       </body>
     </html>
   );
